完成了模板格式。接下来需要增加代理ip搜索和请求时间间隔设置

+++ src/main/java/com/zql/dto/ElementDto.java
@@ -19,7 +19,7 @@ public class ElementDto {
 
     public static class AppMsgExtInfo{
         @SerializedName("audio_fileid")
-        public Integer audioFileid;
+        public Long audioFileid;
         public String author;
         public String content;
         @SerializedName("content_url")
